Don't try to delete forms when there is not one to delete (#1795)

+++ collect_app/src/main/java/org/odk/collect/android/tasks/DiskSyncTask.java
@@ -138,8 +138,10 @@ public class DiskSyncTask extends AsyncTask<Void, String, String> {
                     }
                 }
 
-                //Delete the forms not found in sdcard from the database
-                formsDao.deleteFormsFromIDs(idsToDelete.toArray(new String[idsToDelete.size()]));
+                if (!idsToDelete.isEmpty()) {
+                    //Delete the forms not found in sdcard from the database
+                    formsDao.deleteFormsFromIDs(idsToDelete.toArray(new String[idsToDelete.size()]));
+                }
 
                 // Step3: go through uriToUpdate to parse and update each in turn.
                 // This is slow because buildContentValues(...) is slow.
